+++ CogniteSdk.Types/Relationships/Relationship.cs
@@ -37,7 +37,7 @@ namespace CogniteSdk
         /// existence of the relationship. Generated relationships
         /// should never have the a confidence score of 1.0.
         /// </summary>
-        public float Confidence { get; set; }
+        public double Confidence { get; set; }
 
         /// <summary>
         /// String describing the source system storing or generating the relationship.
